requirements_1.txt
alembic==0.9.2
appnope==0.1.0
asn1crypto==0.24.0
attrs==17.4.0
autobahn==17.10.1
Automat==0.6.0
bleach==1.5.0
blinker==1.4
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
click==6.7
constantly==15.1.0
cryptography==2.1.4
dateparser==0.7.0
decorator==4.0.11
dominate==2.3.1
entrypoints==0.2.2
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Mail==0.9.1
Flask-Migrate==2.0.4
Flask-Moment==0.5.1
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
Flask-WTF==0.14.2
html5lib==0.999
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
itsdangerous==0.24
jedi==0.10.2
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.0.1
jupyter-console==5.1.0
jupyter-core==4.3.0
Mako==1.0.6
MarkupSafe==0.23
mistune==0.7.4
nbconvert==5.2.1
nbformat==4.3.0
notebook==5.0.0
pandocfilters==1.4.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
ptyprocess==0.5.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
Pygments==2.2.0
PyMySQL==0.7.9
pyOpenSSL==17.5.0
python-binance==0.6.6
python-dateutil==2.6.0
python-editor==1.0.3
pytz==2018.3
pyzmq==16.0.2
qtconsole==4.3.0
regex==2018.2.8
requests==2.18.4
service-identity==17.0.0
simplegeneric==0.8.1
six==1.10.0
SQLAlchemy==1.1.11
terminado==0.6
testpath==0.3
tornado==4.5.1
traitlets==4.3.2
Twisted==17.9.0
txaio==2.8.2
tzlocal==1.5.1
urllib3==1.22
visitor==0.1.3
wcwidth==0.1.7
Werkzeug==0.12.2
widgetsnbextension==2.0.0
WTForms==2.1
zope.interface==4.4.3
